requirements.txt
redis
playwright
requests
paramiko
pyyaml